test: split session user idempotency regressions
